% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/praise.R
\name{praise}
\alias{praise}
\title{Title}
\usage{
praise(name, punctuation = "!")
}
\arguments{
\item{punctuation}{text string: This is the emphasis as a text input}

\item{text}{string: The name of the person you want to ensure they are the guy}
}
\value{
text string with praise
}
\description{
this function is for when you want to be that guy, pal
}
\examples{
praise(name = "Mike", punctuation = "!")
}
